+++ src/env.d.ts
@@ -8,7 +8,6 @@ interface ImportMetaEnv {
   readonly VITE_RC_APPKEY: string
   readonly VITE_API_BASE: string
   readonly VITE_OSS_BASE: string
-  readonly VITE_READY_TOKEN: string
 }
 interface ImportMeta {
   readonly env: ImportMetaEnv